revenue prediction is working
